Add multipage support Frontend
Added foundation to enable multiple pages
