+++ Frontend/src/Pages/Profile.js
@@ -1,0 +1,11 @@
+
+import React from "react";
+
+
+function Profile() {
+    return (
+        <h1>Profile page</h1>
+    );
+}
+
+export default Profile;
