+++ Frontend/src/index.js
@@ -1,15 +1,23 @@
 import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import ReactDOM from 'react-dom';
+import './Styles/index.css';
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <App />
+import { Routes, Route } from 'react-router-dom';
+
+import Home from './Pages/Home';
+
+import Profile from './Pages/Profile'
+
+import { BrowserRouter as Router } from 'react-router-dom';
+
+ReactDOM.render(
+  <React.StrictMode>
+    <Router>
+      <Routes>
+        <Route path='/' element={<Home />} />
+        <Route path='/profile' element={<Profile />} />
+      </Routes>
+    </Router>
+  </React.StrictMode>,
+  document.getElementById('root')
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
